Make runonce logging more clear
* Extend the APscheduler __event_listener callback to iterate over
the remaining scheduled jobs after executing a runonce job so that it
will log the next, regularly scheduled job.  Previously, the last log
written after a runonce job was confusing as the APScheduler lib
simply logged when the LAST job was run and did not include when
the next job was scheduled to run.