Make runonce logging more clear (#2)
* Extend the APscheduler __event_listener callback to iterate over
the remaining scheduled jobs after executing a runonce job so that it
will log the next, regularly scheduled job.  Previously, the last log
written after a runonce job was confusing as the APScheduler lib
simply logged when the LAST job was run and did not include when
the next job was scheduled to run.

* Do not fail the test if we attempt to kill a process that
exits on its own after we get the pid and before we can kill
it.